# express_mongo_auth/src/config/swagger/swaggerConfig.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Shape REST API Documentation",
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Development server"
        },
        {
            "url": "http://18.184.143.238:5000/",
            "description": "Staging EC2 server"
        }
    ],
...
    "paths": {
...
        "/program/list/{id}": {
            "get": {
                "tags": [
                    "Program"
                ],
                "summary": "Get program by ID",
                "description": "Returns a single program",
                "operationId": "getProgramById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the program to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found"
                    }
                }
            }
        },
        "/program/list": {
            "get": {
                "tags": [
                    "Program"
                ],
                "summary": "Get all programs",
                "description": "Returns a list of programs",
                "operationId": "getAllPrograms",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Program"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Program"
                ],
                "summary": "Update program by ID",
                "description": "Updates a single program",
                "operationId": "updateProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the program to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Program"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Program"
                ],
                "summary": "Delete program by ID",
                "description": "Deletes a single program",
                "operationId": "deleteProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the program to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Program not found"
                    }
                }
            }
        },
...
        "/program/mentor/list/{programId}": {
            "get": {
                "tags": [
                    "Program"
                ],
                "summary": "Get Active Shapers of a Mentor by of a Specified Program",
                "description": "Returns the shapers and the program",
                "operationId": "getActiveShapersByProgramAndMentor",
                "parameters": [
                    {
                        "name": "programId",
                        "in": "path",
                        "description": "ID of the program",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "program": {
                                            "$ref": "#/components/schemas/Program"
                                        },
                                        "shapers": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Client": {
                "title": "Client",
                "type": "object",
                "properties": {
                    "clientId": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "softSkills": {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "skill": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "skill"
                                    ]
                                }
                            }
                        },
                        "required": [
                            "data"
                        ]
                    },
                    "skillsets": {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "skill": {
                                            "type": "string"
                                        },
                                        "level": {
                                            "type": "string"
                                        },
                                        "isSoftSkill": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "skill",
                                        "level",
                                        "isSoftSkill"
                                    ]
                                }
                            }
                        },
                        "required": [
                            "data"
                        ]
                    },
                    "studies": {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "local": {
                                            "type": "string"
                                        },
                                        "specialty": {
                                            "type": "string"
                                        },
                                        "degree": {
                                            "type": "string"
                                        },
                                        "startDate": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "endDate": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "establishment": {
                                            "type": "string"
                                        },
                                        "student": {
                                            "type": "boolean"
                                        },
                                        "document": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "local",
                                        "specialty",
                                        "degree",
                                        "startDate",
                                        "endDate",
                                        "establishment",
                                        "student"
                                    ]
                                }
                            }
                        },
                        "required": [
                            "data"
                        ]
                    },
                    "professional_experience": {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "company": {
                                            "type": "string"
                                        },
                                        "job": {
                                            "type": "string"
                                        },
                                        "local": {
                                            "type": "string"
                                        },
                                        "period": {
                                            "type": "string"
                                        },
                                        "document": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "company",
                                        "job",
                                        "local",
                                        "period"
                                    ]
                                }
                            }
                        },
                        "required": [
                            "data"
                        ]
                    }
                },
                "required": [
                    "clientId",
                    "softSkills",
                    "skillsets",
                    "studies",
                    "professional_experience"
                ]
            },
...
            "Program": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "published",
                            "archived",
                            "unpublished"
                        ]
                    },
                    "weeks": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "start": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "end": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "lessons": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "$/../../../models/programModels/lesson.ts"
                                    }
                                }
                            },
                            "required": [
                                "start",
                                "end",
                                "lessons"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}